Add link to deploy to Google Cloud
Because it is possible for developers to self-host this service, this change adds guidance about how to deploy it to Google Cloud, by using [Google Cloud Run](https://cloud.run). The fact that this container listens for requests on `$PORT` makes it able to run without any change on the platform.